refactor(table-catalog): split REST handler modules
