+++ rustfs/src/admin/handlers/table_catalog/mod.rs
@@ -45,6 +45,24 @@ use std::time::{Duration as StdDuration, Instant};
 use time::{Duration, OffsetDateTime};
 use uuid::Uuid;
 
+mod config;
+mod credentials;
+mod maintenance;
+mod namespace;
+mod refs;
+mod routes;
+mod table;
+mod view;
+
+use config::*;
+use credentials::*;
+use maintenance::*;
+use namespace::*;
+use refs::*;
+pub use routes::register_table_catalog_route;
+use table::*;
+use view::*;
+
 const JSON_CONTENT_TYPE: &str = "application/json";
 const ENV_TABLE_CATALOG_CREDENTIAL_VENDING: &str = "RUSTFS_TABLE_CATALOG_CREDENTIAL_VENDING";
 const ENV_TABLE_CATALOG_CREDENTIAL_TTL_SECONDS: &str = "RUSTFS_TABLE_CATALOG_CREDENTIAL_TTL_SECONDS";
@@ -865,250 +883,6 @@ struct TableMetadataLocationResponse {
     generation: u64,
     #[serde(rename = "warehouse-location")]
     warehouse_location: String,
-}
-
-pub fn register_table_catalog_route(r: &mut S3Router<AdminOperation>) -> std::io::Result<()> {
-    for prefix in [TABLE_CATALOG_PREFIX, TABLE_CATALOG_COMPAT_PREFIX] {
-        register_table_catalog_prefix_routes(r, prefix)?;
-    }
-
-    Ok(())
-}
-
-fn register_table_catalog_prefix_routes(r: &mut S3Router<AdminOperation>, prefix: &str) -> std::io::Result<()> {
-    r.insert(Method::GET, format!("{prefix}/config").as_str(), AdminOperation(&GET_CONFIG_HANDLER))?;
-    r.insert(
-        Method::PUT,
-        format!("{prefix}/buckets/{{warehouse}}").as_str(),
-        AdminOperation(&ENABLE_TABLE_BUCKET_HANDLER),
-    )?;
-    r.insert(
-        Method::GET,
-        format!("{prefix}/buckets/{{warehouse}}").as_str(),
-        AdminOperation(&GET_TABLE_BUCKET_HANDLER),
-    )?;
-    r.insert(
-        Method::GET,
-        format!("{prefix}/{{warehouse}}/catalog/migration").as_str(),
-        AdminOperation(&GET_TABLE_CATALOG_MIGRATION_HANDLER),
-    )?;
-    r.insert(
-        Method::POST,
-        format!("{prefix}/{{warehouse}}/catalog/migration").as_str(),
-        AdminOperation(&MATERIALIZE_TABLE_CATALOG_MIGRATION_HANDLER),
-    )?;
-    r.insert(
-        Method::DELETE,
-        format!("{prefix}/{{warehouse}}/catalog/migration").as_str(),
-        AdminOperation(&CANCEL_TABLE_CATALOG_MIGRATION_HANDLER),
-    )?;
-    r.insert(
-        Method::GET,
-        format!("{prefix}/{{warehouse}}/namespaces").as_str(),
-        AdminOperation(&LIST_NAMESPACES_HANDLER),
-    )?;
-    r.insert(
-        Method::POST,
-        format!("{prefix}/{{warehouse}}/namespaces").as_str(),
-        AdminOperation(&CREATE_NAMESPACE_HANDLER),
-    )?;
-    r.insert(
-        Method::GET,
-        format!("{prefix}/{{warehouse}}/namespaces/{{namespace}}").as_str(),
-        AdminOperation(&GET_NAMESPACE_HANDLER),
-    )?;
-    r.insert(
-        Method::HEAD,
-        format!("{prefix}/{{warehouse}}/namespaces/{{namespace}}").as_str(),
-        AdminOperation(&NAMESPACE_EXISTS_HANDLER),
-    )?;
-    r.insert(
-        Method::DELETE,
-        format!("{prefix}/{{warehouse}}/namespaces/{{namespace}}").as_str(),
-        AdminOperation(&DROP_NAMESPACE_HANDLER),
-    )?;
-    r.insert(
-        Method::GET,
-        format!("{prefix}/{{warehouse}}/namespaces/{{namespace}}/tables").as_str(),
-        AdminOperation(&LIST_TABLES_HANDLER),
-    )?;
-    r.insert(
-        Method::POST,
-        format!("{prefix}/{{warehouse}}/namespaces/{{namespace}}/tables").as_str(),
-        AdminOperation(&CREATE_TABLE_HANDLER),
-    )?;
-    r.insert(
-        Method::POST,
-        format!("{prefix}/{{warehouse}}/namespaces/{{namespace}}/register").as_str(),
-        AdminOperation(&REGISTER_TABLE_HANDLER),
-    )?;
-    r.insert(
-        Method::GET,
-        format!("{prefix}/{{warehouse}}/namespaces/{{namespace}}/views").as_str(),
-        AdminOperation(&LIST_VIEWS_HANDLER),
-    )?;
-    r.insert(
-        Method::POST,
-        format!("{prefix}/{{warehouse}}/namespaces/{{namespace}}/views").as_str(),
-        AdminOperation(&CREATE_VIEW_HANDLER),
-    )?;
-    r.insert(
-        Method::GET,
-        format!("{prefix}/{{warehouse}}/namespaces/{{namespace}}/tables/{{table}}").as_str(),
-        AdminOperation(&LOAD_TABLE_HANDLER),
-    )?;
-    r.insert(
-        Method::HEAD,
-        format!("{prefix}/{{warehouse}}/namespaces/{{namespace}}/tables/{{table}}").as_str(),
-        AdminOperation(&TABLE_EXISTS_HANDLER),
-    )?;
-    r.insert(
-        Method::GET,
-        format!("{prefix}/{{warehouse}}/namespaces/{{namespace}}/tables/{{table}}/credentials").as_str(),
-        AdminOperation(&LOAD_CREDENTIALS_HANDLER),
-    )?;
-    r.insert(
-        Method::POST,
-        format!("{prefix}/{{warehouse}}/namespaces/{{namespace}}/tables/{{table}}").as_str(),
-        AdminOperation(&COMMIT_TABLE_HANDLER),
-    )?;
-    r.insert(
-        Method::DELETE,
-        format!("{prefix}/{{warehouse}}/namespaces/{{namespace}}/tables/{{table}}").as_str(),
-        AdminOperation(&DROP_TABLE_HANDLER),
-    )?;
-    r.insert(
-        Method::GET,
-        format!("{prefix}/{{warehouse}}/namespaces/{{namespace}}/views/{{view}}").as_str(),
-        AdminOperation(&LOAD_VIEW_HANDLER),
-    )?;
-    r.insert(
-        Method::HEAD,
-        format!("{prefix}/{{warehouse}}/namespaces/{{namespace}}/views/{{view}}").as_str(),
-        AdminOperation(&VIEW_EXISTS_HANDLER),
-    )?;
-    r.insert(
-        Method::POST,
-        format!("{prefix}/{{warehouse}}/namespaces/{{namespace}}/views/{{view}}").as_str(),
-        AdminOperation(&REPLACE_VIEW_HANDLER),
-    )?;
-    r.insert(
-        Method::DELETE,
-        format!("{prefix}/{{warehouse}}/namespaces/{{namespace}}/views/{{view}}").as_str(),
-        AdminOperation(&DROP_VIEW_HANDLER),
-    )?;
-    r.insert(
-        Method::GET,
-        format!("{prefix}/{{warehouse}}/namespaces/{{namespace}}/tables/{{table}}/refs").as_str(),
-        AdminOperation(&LIST_TABLE_REFS_HANDLER),
-    )?;
-    r.insert(
-        Method::PUT,
-        format!("{prefix}/{{warehouse}}/namespaces/{{namespace}}/tables/{{table}}/refs/{{ref}}").as_str(),
-        AdminOperation(&PUT_TABLE_REF_HANDLER),
-    )?;
-    r.insert(
-        Method::DELETE,
-        format!("{prefix}/{{warehouse}}/namespaces/{{namespace}}/tables/{{table}}/refs/{{ref}}").as_str(),
-        AdminOperation(&DELETE_TABLE_REF_HANDLER),
-    )?;
-    r.insert(
-        Method::GET,
-        format!("{prefix}/{{warehouse}}/namespaces/{{namespace}}/tables/{{table}}/metadata-location").as_str(),
-        AdminOperation(&GET_TABLE_METADATA_LOCATION_HANDLER),
-    )?;
-    r.insert(
-        Method::PUT,
-        format!("{prefix}/{{warehouse}}/namespaces/{{namespace}}/tables/{{table}}/metadata-location").as_str(),
-        AdminOperation(&UPDATE_TABLE_METADATA_LOCATION_HANDLER),
-    )?;
-    r.insert(
-        Method::POST,
-        format!("{prefix}/{{warehouse}}/namespaces/{{namespace}}/tables/{{table}}/maintenance/metadata").as_str(),
-        AdminOperation(&TABLE_METADATA_MAINTENANCE_HANDLER),
-    )?;
-    r.insert(
-        Method::GET,
-        format!("{prefix}/{{warehouse}}/namespaces/{{namespace}}/tables/{{table}}/maintenance/config").as_str(),
-        AdminOperation(&GET_TABLE_MAINTENANCE_CONFIG_HANDLER),
-    )?;
-    r.insert(
-        Method::PUT,
-        format!("{prefix}/{{warehouse}}/namespaces/{{namespace}}/tables/{{table}}/maintenance/config").as_str(),
-        AdminOperation(&PUT_TABLE_MAINTENANCE_CONFIG_HANDLER),
-    )?;
-    r.insert(
-        Method::GET,
-        format!("{prefix}/{{warehouse}}/namespaces/{{namespace}}/tables/{{table}}/maintenance/jobs/{{job}}").as_str(),
-        AdminOperation(&GET_TABLE_MAINTENANCE_JOB_HANDLER),
-    )?;
-    r.insert(
-        Method::GET,
-        format!("{prefix}/{{warehouse}}/namespaces/{{namespace}}/tables/{{table}}/maintenance/scheduler").as_str(),
-        AdminOperation(&GET_TABLE_MAINTENANCE_SCHEDULER_HANDLER),
-    )?;
-    r.insert(
-        Method::POST,
-        format!("{prefix}/{{warehouse}}/namespaces/{{namespace}}/tables/{{table}}/maintenance/scheduler/run").as_str(),
-        AdminOperation(&RUN_TABLE_MAINTENANCE_SCHEDULER_HANDLER),
-    )?;
-    r.insert(
-        Method::POST,
-        format!("{prefix}/{{warehouse}}/namespaces/{{namespace}}/tables/{{table}}/maintenance/worker/run").as_str(),
-        AdminOperation(&RUN_TABLE_MAINTENANCE_WORKER_HANDLER),
-    )?;
-    r.insert(
-        Method::POST,
-        format!("{prefix}/{{warehouse}}/namespaces/{{namespace}}/tables/{{table}}/maintenance/jobs/{{job}}/heartbeat").as_str(),
-        AdminOperation(&HEARTBEAT_TABLE_MAINTENANCE_JOB_HANDLER),
-    )?;
-    r.insert(
-        Method::POST,
-        format!("{prefix}/{{warehouse}}/namespaces/{{namespace}}/tables/{{table}}/maintenance/jobs/{{job}}/quarantine").as_str(),
-        AdminOperation(&TABLE_MAINTENANCE_QUARANTINE_HANDLER),
-    )?;
-    r.insert(
-        Method::GET,
-        format!("{prefix}/{{warehouse}}/namespaces/{{namespace}}/tables/{{table}}/catalog/export").as_str(),
-        AdminOperation(&EXPORT_TABLE_CATALOG_HANDLER),
-    )?;
-    r.insert(
-        Method::POST,
-        format!("{prefix}/{{warehouse}}/namespaces/{{namespace}}/tables/{{table}}/catalog/import").as_str(),
-        AdminOperation(&IMPORT_TABLE_CATALOG_HANDLER),
-    )?;
-    r.insert(
-        Method::GET,
-        format!("{prefix}/{{warehouse}}/namespaces/{{namespace}}/tables/{{table}}/catalog/external").as_str(),
-        AdminOperation(&EXTERNAL_CATALOG_BRIDGE_HANDLER),
-    )?;
-    r.insert(
-        Method::PUT,
-        format!("{prefix}/{{warehouse}}/namespaces/{{namespace}}/tables/{{table}}/catalog/external").as_str(),
-        AdminOperation(&PUT_EXTERNAL_CATALOG_BRIDGE_HANDLER),
-    )?;
-    r.insert(
-        Method::POST,
-        format!("{prefix}/{{warehouse}}/namespaces/{{namespace}}/tables/{{table}}/catalog/external/sync").as_str(),
-        AdminOperation(&SYNC_EXTERNAL_CATALOG_BRIDGE_HANDLER),
-    )?;
-    r.insert(
-        Method::GET,
-        format!("{prefix}/{{warehouse}}/namespaces/{{namespace}}/tables/{{table}}/catalog/diagnostics").as_str(),
-        AdminOperation(&GET_TABLE_CATALOG_DIAGNOSTICS_HANDLER),
-    )?;
-    r.insert(
-        Method::POST,
-        format!("{prefix}/{{warehouse}}/namespaces/{{namespace}}/tables/{{table}}/catalog/recovery").as_str(),
-        AdminOperation(&RECOVER_TABLE_CATALOG_HANDLER),
-    )?;
-    r.insert(
-        Method::POST,
-        format!("{prefix}/{{warehouse}}/namespaces/{{namespace}}/tables/{{table}}/catalog/rollback").as_str(),
-        AdminOperation(&ROLLBACK_TABLE_CATALOG_HANDLER),
-    )?;
-
-    Ok(())
 }
 
 fn catalog_config_response(warehouse: Option<&str>) -> S3Result<CatalogConfigResponse> {
@@ -5217,930 +4991,6 @@ where
     .await;
     record_table_catalog_admin_operation_result("rollback", bucket, &namespace.public_name(), table, started, &result);
     result
-}
-
-pub struct GetCatalogConfigHandler {}
-
-#[async_trait::async_trait]
-impl Operation for GetCatalogConfigHandler {
-    async fn call(&self, req: S3Request<Body>, _params: Params<'_, '_>) -> S3Result<S3Response<(StatusCode, Body)>> {
-        authorize_table_catalog_request(&req, AdminAction::GetTableCatalogAction).await?;
-        let warehouse = warehouse_from_config_query(&req.uri)?;
-        build_json_response(StatusCode::OK, &catalog_config_response(warehouse.as_deref())?)
-    }
-}
-
-pub struct EnableTableBucketHandler {}
-
-#[async_trait::async_trait]
-impl Operation for EnableTableBucketHandler {
-    async fn call(&self, req: S3Request<Body>, params: Params<'_, '_>) -> S3Result<S3Response<(StatusCode, Body)>> {
-        let warehouse = warehouse_from_params(&params)?;
-        let resource = TableCatalogResource::warehouse(&warehouse);
-        authorize_table_catalog_resource_request(&req, &resource, AdminAction::SetTableBucketAction).await?;
-        let store = table_catalog_store()?;
-        let response = enable_table_bucket_response(&store, &warehouse).await?;
-        build_json_response(StatusCode::OK, &response)
-    }
-}
-
-pub struct GetTableBucketHandler {}
-
-#[async_trait::async_trait]
-impl Operation for GetTableBucketHandler {
-    async fn call(&self, req: S3Request<Body>, params: Params<'_, '_>) -> S3Result<S3Response<(StatusCode, Body)>> {
-        let warehouse = warehouse_from_params(&params)?;
-        let resource = TableCatalogResource::warehouse(&warehouse);
-        authorize_table_catalog_resource_request(&req, &resource, AdminAction::GetTableBucketAction).await?;
-        let store = table_catalog_store()?;
-        let enabled = table_bucket_enabled_from_metadata(&warehouse).await?;
-        let response = table_bucket_response(&store, &warehouse, enabled).await?;
-        build_json_response(StatusCode::OK, &response)
-    }
-}
-
-pub struct GetTableCatalogMigrationHandler {}
-
-#[async_trait::async_trait]
-impl Operation for GetTableCatalogMigrationHandler {
-    async fn call(&self, req: S3Request<Body>, params: Params<'_, '_>) -> S3Result<S3Response<(StatusCode, Body)>> {
-        let warehouse = warehouse_from_params(&params)?;
-        let resource = TableCatalogResource::warehouse(&warehouse);
-        authorize_table_catalog_resource_request(&req, &resource, AdminAction::GetTableCatalogAction).await?;
-        ensure_table_bucket_enabled(&warehouse).await?;
-        let store = table_catalog_object_store()?;
-        let started = Instant::now();
-        let result = store
-            .plan_durable_strong_backing_migration(&warehouse)
-            .await
-            .map_err(catalog_store_error);
-        record_table_catalog_admin_operation_result("migration", &warehouse, "", "", started, &result);
-        let response = result?;
-        build_json_response(StatusCode::OK, &response)
-    }
-}
-
-pub struct MaterializeTableCatalogMigrationHandler {}
-
-#[async_trait::async_trait]
-impl Operation for MaterializeTableCatalogMigrationHandler {
-    async fn call(&self, req: S3Request<Body>, params: Params<'_, '_>) -> S3Result<S3Response<(StatusCode, Body)>> {
-        let warehouse = warehouse_from_params(&params)?;
-        authorize_table_catalog_request(&req, AdminAction::MigrateTableCatalogAction).await?;
-        ensure_table_bucket_enabled(&warehouse).await?;
-        let store = table_catalog_object_store()?;
-        let started = Instant::now();
-        let result = store
-            .materialize_durable_strong_backing_migration(&warehouse)
-            .await
-            .map_err(catalog_store_error);
-        record_table_catalog_admin_operation_result("migration-materialize", &warehouse, "", "", started, &result);
-        let response = result?;
-        build_json_response(StatusCode::OK, &response)
-    }
-}
-
-pub struct CancelTableCatalogMigrationHandler {}
-
-#[async_trait::async_trait]
-impl Operation for CancelTableCatalogMigrationHandler {
-    async fn call(&self, req: S3Request<Body>, params: Params<'_, '_>) -> S3Result<S3Response<(StatusCode, Body)>> {
-        let warehouse = warehouse_from_params(&params)?;
-        authorize_table_catalog_request(&req, AdminAction::MigrateTableCatalogAction).await?;
-        ensure_table_bucket_enabled(&warehouse).await?;
-        let store = table_catalog_object_store()?;
-        let started = Instant::now();
-        let result = store
-            .cancel_durable_strong_backing_migration(&warehouse)
-            .await
-            .map_err(catalog_store_error);
-        record_table_catalog_admin_operation_result("migration-cancel", &warehouse, "", "", started, &result);
-        let response = result?;
-        build_json_response(StatusCode::OK, &response)
-    }
-}
-
-pub struct RestListNamespacesHandler {}
-
-#[async_trait::async_trait]
-impl Operation for RestListNamespacesHandler {
-    async fn call(&self, req: S3Request<Body>, params: Params<'_, '_>) -> S3Result<S3Response<(StatusCode, Body)>> {
-        let warehouse = warehouse_from_params(&params)?;
-        let resource = TableCatalogResource::warehouse(&warehouse);
-        authorize_table_catalog_resource_request(&req, &resource, AdminAction::GetTableNamespaceAction).await?;
-        ensure_table_bucket_enabled(&warehouse).await?;
-        let store = table_catalog_store()?;
-        let response = list_namespaces_response(&store, &warehouse, &req.uri).await?;
-        build_json_response(StatusCode::OK, &response)
-    }
-}
-
-pub struct RestCreateNamespaceHandler {}
-
-#[async_trait::async_trait]
-impl Operation for RestCreateNamespaceHandler {
-    async fn call(&self, req: S3Request<Body>, params: Params<'_, '_>) -> S3Result<S3Response<(StatusCode, Body)>> {
-        let warehouse = warehouse_from_params(&params)?;
-        let resource = TableCatalogResource::warehouse(&warehouse);
-        authorize_table_catalog_resource_request(&req, &resource, AdminAction::SetTableNamespaceAction).await?;
-        let request = read_json_body::<CreateNamespaceRequest>(req.input).await?;
-        let store = table_catalog_store()?;
-        let table_bucket_enabled = table_bucket_enabled_from_metadata(&warehouse).await?;
-        let response = create_namespace_response(&store, &warehouse, request, table_bucket_enabled).await?;
-        build_json_response(StatusCode::OK, &response)
-    }
-}
-
-pub struct RestGetNamespaceHandler {}
-
-#[async_trait::async_trait]
-impl Operation for RestGetNamespaceHandler {
-    async fn call(&self, req: S3Request<Body>, params: Params<'_, '_>) -> S3Result<S3Response<(StatusCode, Body)>> {
-        let warehouse = warehouse_from_params(&params)?;
-        let namespace = namespace_from_params(&params)?;
-        let resource = TableCatalogResource::namespace(&warehouse, &namespace);
-        authorize_table_catalog_resource_request(&req, &resource, AdminAction::GetTableNamespaceAction).await?;
-        ensure_table_bucket_enabled(&warehouse).await?;
-        let store = table_catalog_store()?;
-        let response = get_namespace_response(&store, &warehouse, &namespace).await?;
-        build_json_response(StatusCode::OK, &response)
-    }
-}
-
-pub struct RestDropNamespaceHandler {}
-
-#[async_trait::async_trait]
-impl Operation for RestDropNamespaceHandler {
-    async fn call(&self, req: S3Request<Body>, params: Params<'_, '_>) -> S3Result<S3Response<(StatusCode, Body)>> {
-        let warehouse = warehouse_from_params(&params)?;
-        let namespace = namespace_from_params(&params)?;
-        let resource = TableCatalogResource::namespace(&warehouse, &namespace);
-        authorize_table_catalog_resource_request(&req, &resource, AdminAction::DeleteTableNamespaceAction).await?;
-        ensure_table_bucket_enabled(&warehouse).await?;
-        let store = table_catalog_store()?;
-        drop_namespace_in_store(&store, &warehouse, &namespace.public_name()).await?;
-        Ok(empty_response(StatusCode::NO_CONTENT))
-    }
-}
-
-pub struct RestNamespaceExistsHandler {}
-
-#[async_trait::async_trait]
-impl Operation for RestNamespaceExistsHandler {
-    async fn call(&self, req: S3Request<Body>, params: Params<'_, '_>) -> S3Result<S3Response<(StatusCode, Body)>> {
-        let warehouse = warehouse_from_params(&params)?;
-        let namespace = namespace_from_params(&params)?;
-        let resource = TableCatalogResource::namespace(&warehouse, &namespace);
-        authorize_table_catalog_resource_request(&req, &resource, AdminAction::GetTableNamespaceAction).await?;
-        ensure_table_bucket_enabled(&warehouse).await?;
-        let store = table_catalog_store()?;
-        Ok(empty_response(namespace_exists_status(&store, &warehouse, &namespace).await?))
-    }
-}
-
-pub struct RestListTablesHandler {}
-
-#[async_trait::async_trait]
-impl Operation for RestListTablesHandler {
-    async fn call(&self, req: S3Request<Body>, params: Params<'_, '_>) -> S3Result<S3Response<(StatusCode, Body)>> {
-        let warehouse = warehouse_from_params(&params)?;
-        let namespace = namespace_from_params(&params)?;
-        let resource = TableCatalogResource::namespace(&warehouse, &namespace);
-        authorize_table_catalog_resource_request(&req, &resource, AdminAction::GetTableAction).await?;
-        ensure_table_bucket_enabled(&warehouse).await?;
-        let store = table_catalog_store()?;
-        let response = list_tables_response(&store, &warehouse, &namespace, &req.uri).await?;
-        build_json_response(StatusCode::OK, &response)
-    }
-}
-
-pub struct RestCreateTableHandler {}
-
-#[async_trait::async_trait]
-impl Operation for RestCreateTableHandler {
-    async fn call(&self, req: S3Request<Body>, params: Params<'_, '_>) -> S3Result<S3Response<(StatusCode, Body)>> {
-        let warehouse = warehouse_from_params(&params)?;
-        let namespace = namespace_from_params(&params)?;
-        let resource = TableCatalogResource::namespace(&warehouse, &namespace);
-        authorize_table_catalog_resource_request(&req, &resource, AdminAction::CreateTableAction).await?;
-        let request = read_json_body::<CreateTableRequest>(req.input).await?;
-        let metadata_backend = table_catalog_backend()?;
-        let store = table_catalog_store_from_backend(metadata_backend.clone())?;
-        let table_bucket_enabled = table_bucket_enabled_from_metadata(&warehouse).await?;
-        let response =
-            create_table_response(&store, &metadata_backend, &warehouse, &namespace, request, table_bucket_enabled).await?;
-        build_json_response(StatusCode::OK, &response)
-    }
-}
-
-pub struct RestRegisterTableHandler {}
-
-#[async_trait::async_trait]
-impl Operation for RestRegisterTableHandler {
-    async fn call(&self, req: S3Request<Body>, params: Params<'_, '_>) -> S3Result<S3Response<(StatusCode, Body)>> {
-        let warehouse = warehouse_from_params(&params)?;
-        let namespace = namespace_from_params(&params)?;
-        let resource = TableCatalogResource::namespace(&warehouse, &namespace);
-        authorize_table_catalog_resource_request(&req, &resource, AdminAction::RegisterTableAction).await?;
-        let request = read_json_body::<RegisterTableRequest>(req.input).await?;
-        let metadata_backend = table_catalog_backend()?;
-        let store = table_catalog_store_from_backend(metadata_backend.clone())?;
-        let table_bucket_enabled = table_bucket_enabled_from_metadata(&warehouse).await?;
-        let response =
-            register_table_response(&store, &metadata_backend, &warehouse, &namespace, request, table_bucket_enabled).await?;
-        build_json_response(StatusCode::OK, &response)
-    }
-}
-
-pub struct RestListViewsHandler {}
-
-#[async_trait::async_trait]
-impl Operation for RestListViewsHandler {
-    async fn call(&self, req: S3Request<Body>, params: Params<'_, '_>) -> S3Result<S3Response<(StatusCode, Body)>> {
-        let warehouse = warehouse_from_params(&params)?;
-        let namespace = namespace_from_params(&params)?;
-        let resource = TableCatalogResource::namespace(&warehouse, &namespace);
-        authorize_table_catalog_resource_request(&req, &resource, AdminAction::GetTableMetadataAction).await?;
-        ensure_table_bucket_enabled(&warehouse).await?;
-        let store = table_catalog_store()?;
-        let response = list_views_response(&store, &warehouse, &namespace, &req.uri).await?;
-        build_json_response(StatusCode::OK, &response)
-    }
-}
-
-pub struct RestCreateViewHandler {}
-
-#[async_trait::async_trait]
-impl Operation for RestCreateViewHandler {
-    async fn call(&self, req: S3Request<Body>, params: Params<'_, '_>) -> S3Result<S3Response<(StatusCode, Body)>> {
-        let warehouse = warehouse_from_params(&params)?;
-        let namespace = namespace_from_params(&params)?;
-        let resource = TableCatalogResource::namespace(&warehouse, &namespace);
-        authorize_table_catalog_resource_request(&req, &resource, AdminAction::CreateTableAction).await?;
-        let request = read_json_body::<CreateViewRequest>(req.input).await?;
-        let metadata_backend = table_catalog_backend()?;
-        let store = table_catalog_store_from_backend(metadata_backend.clone())?;
-        let table_bucket_enabled = table_bucket_enabled_from_metadata(&warehouse).await?;
-        let response =
-            create_view_response(&store, &metadata_backend, &warehouse, &namespace, request, table_bucket_enabled).await?;
-        build_json_response(StatusCode::OK, &response)
-    }
-}
-
-pub struct RestLoadTableHandler {}
-
-#[async_trait::async_trait]
-impl Operation for RestLoadTableHandler {
-    async fn call(&self, req: S3Request<Body>, params: Params<'_, '_>) -> S3Result<S3Response<(StatusCode, Body)>> {
-        let warehouse = warehouse_from_params(&params)?;
-        let namespace = namespace_from_params(&params)?;
-        let table = table_name_from_params(&params)?;
-        let resource = TableCatalogResource::table(&warehouse, &namespace, &table);
-        authorize_table_catalog_resource_request(&req, &resource, AdminAction::GetTableMetadataAction).await?;
-        ensure_table_bucket_enabled(&warehouse).await?;
-        let metadata_backend = table_catalog_backend()?;
-        let store = table_catalog_store_from_backend(metadata_backend.clone())?;
-        let response = load_table_response(&store, &metadata_backend, &warehouse, &namespace, &table).await?;
-        build_json_response(StatusCode::OK, &response)
-    }
-}
-
-pub struct RestTableExistsHandler {}
-
-#[async_trait::async_trait]
-impl Operation for RestTableExistsHandler {
-    async fn call(&self, req: S3Request<Body>, params: Params<'_, '_>) -> S3Result<S3Response<(StatusCode, Body)>> {
-        let warehouse = warehouse_from_params(&params)?;
-        let namespace = namespace_from_params(&params)?;
-        let table = table_name_from_params(&params)?;
-        let resource = TableCatalogResource::table(&warehouse, &namespace, &table);
-        authorize_table_catalog_resource_request(&req, &resource, AdminAction::GetTableAction).await?;
-        ensure_table_bucket_enabled(&warehouse).await?;
-        let store = table_catalog_store()?;
-        Ok(empty_response(table_exists_status(&store, &warehouse, &namespace, &table).await?))
-    }
-}
-
-pub struct RestLoadCredentialsHandler {}
-
-#[async_trait::async_trait]
-impl Operation for RestLoadCredentialsHandler {
-    async fn call(&self, req: S3Request<Body>, params: Params<'_, '_>) -> S3Result<S3Response<(StatusCode, Body)>> {
-        let warehouse = warehouse_from_params(&params)?;
-        let namespace = namespace_from_params(&params)?;
-        let table = table_name_from_params(&params)?;
-        let resource = TableCatalogResource::table(&warehouse, &namespace, &table);
-        authorize_table_catalog_resource_request(&req, &resource, AdminAction::GetTableCredentialsAction).await?;
-        ensure_table_bucket_enabled(&warehouse).await?;
-        let principal = table_catalog_request_principal(&req).await?;
-        let store = table_catalog_store()?;
-        let issuer = IamTableCredentialIssuer::from_env();
-        let response = load_credentials_response(&store, &warehouse, &namespace, &table, &issuer, Some(&principal)).await?;
-        build_json_response(StatusCode::OK, &response)
-    }
-}
-
-pub struct RestCommitTableHandler {}
-
-#[async_trait::async_trait]
-impl Operation for RestCommitTableHandler {
-    async fn call(&self, req: S3Request<Body>, params: Params<'_, '_>) -> S3Result<S3Response<(StatusCode, Body)>> {
-        let warehouse = warehouse_from_params(&params)?;
-        let namespace = namespace_from_params(&params)?;
-        let table = table_name_from_params(&params)?;
-        let resource = TableCatalogResource::table(&warehouse, &namespace, &table);
-        authorize_table_catalog_resource_request(&req, &resource, AdminAction::CommitTableAction).await?;
-        ensure_table_bucket_enabled(&warehouse).await?;
-        let request = read_json_body::<RestCommitTableRequest>(req.input).await?;
-        let metadata_backend = table_catalog_backend()?;
-        let store = table_catalog_store_from_backend(metadata_backend.clone())?;
-        let response = commit_table_response(&store, &metadata_backend, &warehouse, &namespace, &table, request).await?;
-        build_json_response(StatusCode::OK, &response)
-    }
-}
-
-pub struct RestDropTableHandler {}
-
-#[async_trait::async_trait]
-impl Operation for RestDropTableHandler {
-    async fn call(&self, req: S3Request<Body>, params: Params<'_, '_>) -> S3Result<S3Response<(StatusCode, Body)>> {
-        let warehouse = warehouse_from_params(&params)?;
-        let namespace = namespace_from_params(&params)?;
-        let table = table_name_from_params(&params)?;
-        let resource = TableCatalogResource::table(&warehouse, &namespace, &table);
-        authorize_table_catalog_resource_request(&req, &resource, AdminAction::DeleteTableAction).await?;
-        ensure_table_bucket_enabled(&warehouse).await?;
-        let store = table_catalog_store()?;
-        drop_table_in_store(&store, &warehouse, &namespace, &table).await?;
-        Ok(empty_response(StatusCode::NO_CONTENT))
-    }
-}
-
-pub struct RestLoadViewHandler {}
-
-#[async_trait::async_trait]
-impl Operation for RestLoadViewHandler {
-    async fn call(&self, req: S3Request<Body>, params: Params<'_, '_>) -> S3Result<S3Response<(StatusCode, Body)>> {
-        let warehouse = warehouse_from_params(&params)?;
-        let namespace = namespace_from_params(&params)?;
-        let view = view_name_from_params(&params)?;
-        let resource = TableCatalogResource::view(&warehouse, &namespace, &view);
-        authorize_table_catalog_resource_request(&req, &resource, AdminAction::GetTableMetadataAction).await?;
-        ensure_table_bucket_enabled(&warehouse).await?;
-        let metadata_backend = table_catalog_backend()?;
-        let store = table_catalog_store_from_backend(metadata_backend.clone())?;
-        let response = load_view_response(&store, &metadata_backend, &warehouse, &namespace, &view).await?;
-        build_json_response(StatusCode::OK, &response)
-    }
-}
-
-pub struct RestViewExistsHandler {}
-
-#[async_trait::async_trait]
-impl Operation for RestViewExistsHandler {
-    async fn call(&self, req: S3Request<Body>, params: Params<'_, '_>) -> S3Result<S3Response<(StatusCode, Body)>> {
-        let warehouse = warehouse_from_params(&params)?;
-        let namespace = namespace_from_params(&params)?;
-        let view = view_name_from_params(&params)?;
-        let resource = TableCatalogResource::view(&warehouse, &namespace, &view);
-        authorize_table_catalog_resource_request(&req, &resource, AdminAction::GetTableAction).await?;
-        ensure_table_bucket_enabled(&warehouse).await?;
-        let store = table_catalog_store()?;
-        Ok(empty_response(view_exists_status(&store, &warehouse, &namespace, &view).await?))
-    }
-}
-
-pub struct RestReplaceViewHandler {}
-
-#[async_trait::async_trait]
-impl Operation for RestReplaceViewHandler {
-    async fn call(&self, req: S3Request<Body>, params: Params<'_, '_>) -> S3Result<S3Response<(StatusCode, Body)>> {
-        let warehouse = warehouse_from_params(&params)?;
-        let namespace = namespace_from_params(&params)?;
-        let view = view_name_from_params(&params)?;
-        let resource = TableCatalogResource::view(&warehouse, &namespace, &view);
-        authorize_table_catalog_resource_request(&req, &resource, AdminAction::CommitTableAction).await?;
-        ensure_table_bucket_enabled(&warehouse).await?;
-        let request = read_json_body::<RestCommitViewRequest>(req.input).await?;
-        let metadata_backend = table_catalog_backend()?;
-        let store = table_catalog_store_from_backend(metadata_backend.clone())?;
-        let response = replace_view_response(&store, &metadata_backend, &warehouse, &namespace, &view, request).await?;
-        build_json_response(StatusCode::OK, &response)
-    }
-}
-
-pub struct RestDropViewHandler {}
-
-#[async_trait::async_trait]
-impl Operation for RestDropViewHandler {
-    async fn call(&self, req: S3Request<Body>, params: Params<'_, '_>) -> S3Result<S3Response<(StatusCode, Body)>> {
-        let warehouse = warehouse_from_params(&params)?;
-        let namespace = namespace_from_params(&params)?;
-        let view = view_name_from_params(&params)?;
-        let resource = TableCatalogResource::view(&warehouse, &namespace, &view);
-        authorize_table_catalog_resource_request(&req, &resource, AdminAction::DeleteTableAction).await?;
-        ensure_table_bucket_enabled(&warehouse).await?;
-        let store = table_catalog_store()?;
-        drop_view_in_store(&store, &warehouse, &namespace, &view).await?;
-        Ok(empty_response(StatusCode::NO_CONTENT))
-    }
-}
-
-pub struct ListTableRefsHandler {}
-
-#[async_trait::async_trait]
-impl Operation for ListTableRefsHandler {
-    async fn call(&self, req: S3Request<Body>, params: Params<'_, '_>) -> S3Result<S3Response<(StatusCode, Body)>> {
-        let warehouse = warehouse_from_params(&params)?;
-        let namespace = namespace_from_params(&params)?;
-        let table = table_name_from_params(&params)?;
-        let resource = TableCatalogResource::table(&warehouse, &namespace, &table);
-        authorize_table_catalog_resource_request(&req, &resource, AdminAction::GetTableMetadataAction).await?;
-        ensure_table_bucket_enabled(&warehouse).await?;
-        let metadata_backend = table_catalog_backend()?;
-        let store = table_catalog_store_from_backend(metadata_backend.clone())?;
-        let response = table_refs_response(&store, &metadata_backend, &warehouse, &namespace, &table).await?;
-        build_json_response(StatusCode::OK, &response)
-    }
-}
-
-pub struct PutTableRefHandler {}
-
-#[async_trait::async_trait]
-impl Operation for PutTableRefHandler {
-    async fn call(&self, req: S3Request<Body>, params: Params<'_, '_>) -> S3Result<S3Response<(StatusCode, Body)>> {
-        let warehouse = warehouse_from_params(&params)?;
-        let namespace = namespace_from_params(&params)?;
-        let table = table_name_from_params(&params)?;
-        let ref_name = ref_name_from_params(&params)?;
-        let resource = TableCatalogResource::table(&warehouse, &namespace, &table);
-        authorize_table_catalog_resource_request(&req, &resource, AdminAction::CommitTableAction).await?;
-        ensure_table_bucket_enabled(&warehouse).await?;
-        let request = read_json_body::<PutTableRefRequest>(req.input).await?;
-        let metadata_backend = table_catalog_backend()?;
-        let store = table_catalog_store_from_backend(metadata_backend.clone())?;
-        let response =
-            put_table_ref_response(&store, &metadata_backend, &warehouse, &namespace, &table, &ref_name, request).await?;
-        build_json_response(StatusCode::OK, &response)
-    }
-}
-
-pub struct DeleteTableRefHandler {}
-
-#[async_trait::async_trait]
-impl Operation for DeleteTableRefHandler {
-    async fn call(&self, req: S3Request<Body>, params: Params<'_, '_>) -> S3Result<S3Response<(StatusCode, Body)>> {
-        let warehouse = warehouse_from_params(&params)?;
-        let namespace = namespace_from_params(&params)?;
-        let table = table_name_from_params(&params)?;
-        let ref_name = ref_name_from_params(&params)?;
-        let resource = TableCatalogResource::table(&warehouse, &namespace, &table);
-        authorize_table_catalog_resource_request(&req, &resource, AdminAction::CommitTableAction).await?;
-        ensure_table_bucket_enabled(&warehouse).await?;
-        let request = read_json_body_or_default::<DeleteTableRefRequest>(req.input).await?;
-        let metadata_backend = table_catalog_backend()?;
-        let store = table_catalog_store_from_backend(metadata_backend.clone())?;
-        let response =
-            delete_table_ref_response(&store, &metadata_backend, &warehouse, &namespace, &table, &ref_name, request).await?;
-        build_json_response(StatusCode::OK, &response)
-    }
-}
-
-pub struct GetTableMetadataLocationHandler {}
-
-#[async_trait::async_trait]
-impl Operation for GetTableMetadataLocationHandler {
-    async fn call(&self, req: S3Request<Body>, params: Params<'_, '_>) -> S3Result<S3Response<(StatusCode, Body)>> {
-        let warehouse = warehouse_from_params(&params)?;
-        let namespace = namespace_from_params(&params)?;
-        let table = table_name_from_params(&params)?;
-        let resource = TableCatalogResource::table(&warehouse, &namespace, &table);
-        authorize_table_catalog_resource_request(&req, &resource, AdminAction::GetTableMetadataLocationAction).await?;
-        ensure_table_bucket_enabled(&warehouse).await?;
-        let store = table_catalog_store()?;
-        let response = get_table_metadata_location_response(&store, &warehouse, &namespace, &table).await?;
-        build_json_response(StatusCode::OK, &response)
-    }
-}
-
-pub struct UpdateTableMetadataLocationHandler {}
-
-#[async_trait::async_trait]
-impl Operation for UpdateTableMetadataLocationHandler {
-    async fn call(&self, req: S3Request<Body>, params: Params<'_, '_>) -> S3Result<S3Response<(StatusCode, Body)>> {
-        let warehouse = warehouse_from_params(&params)?;
-        let namespace = namespace_from_params(&params)?;
-        let table = table_name_from_params(&params)?;
-        let resource = TableCatalogResource::table(&warehouse, &namespace, &table);
-        authorize_table_catalog_resource_request(&req, &resource, AdminAction::SetTableMetadataLocationAction).await?;
-        ensure_table_bucket_enabled(&warehouse).await?;
-        let request = read_json_body::<UpdateTableMetadataLocationRequest>(req.input).await?;
-        let metadata_backend = table_catalog_backend()?;
-        let store = table_catalog_store_from_backend(metadata_backend.clone())?;
-        let response =
-            update_table_metadata_location_response(&store, &metadata_backend, &warehouse, &namespace, &table, request).await?;
-        build_json_response(StatusCode::OK, &response)
-    }
-}
-
-pub struct RestTableMetadataMaintenanceHandler {}
-
-#[async_trait::async_trait]
-impl Operation for RestTableMetadataMaintenanceHandler {
-    async fn call(&self, req: S3Request<Body>, params: Params<'_, '_>) -> S3Result<S3Response<(StatusCode, Body)>> {
-        let warehouse = warehouse_from_params(&params)?;
-        let namespace = namespace_from_params(&params)?;
-        let table = table_name_from_params(&params)?;
-        let resource = TableCatalogResource::table(&warehouse, &namespace, &table);
-        authorize_table_catalog_resource_request(&req, &resource, AdminAction::RunTableMaintenanceAction).await?;
-        ensure_table_bucket_enabled(&warehouse).await?;
-        let request = read_json_body::<TableMetadataMaintenanceRequest>(req.input).await?;
-        let metadata_backend = table_catalog_backend()?;
-        let store = table_catalog_object_store()?;
-        let response =
-            table_metadata_maintenance_response(&store, &metadata_backend, &warehouse, &namespace, &table, request).await?;
-        build_json_response(StatusCode::OK, &response)
-    }
-}
-
-pub struct GetTableMaintenanceConfigHandler {}
-
-#[async_trait::async_trait]
-impl Operation for GetTableMaintenanceConfigHandler {
-    async fn call(&self, req: S3Request<Body>, params: Params<'_, '_>) -> S3Result<S3Response<(StatusCode, Body)>> {
-        let warehouse = warehouse_from_params(&params)?;
-        let namespace = namespace_from_params(&params)?;
-        let table = table_name_from_params(&params)?;
-        let resource = TableCatalogResource::table(&warehouse, &namespace, &table);
-        authorize_table_catalog_resource_request(&req, &resource, AdminAction::GetTableLifecycleAction).await?;
-        ensure_table_bucket_enabled(&warehouse).await?;
-        let store = table_catalog_store()?;
-        let response = store
-            .get_table_maintenance_config(&warehouse, &namespace.public_name(), &table)
-            .await
-            .map_err(catalog_store_error)?;
-        build_json_response(StatusCode::OK, &response)
-    }
-}
-
-pub struct PutTableMaintenanceConfigHandler {}
-
-#[async_trait::async_trait]
-impl Operation for PutTableMaintenanceConfigHandler {
-    async fn call(&self, req: S3Request<Body>, params: Params<'_, '_>) -> S3Result<S3Response<(StatusCode, Body)>> {
-        let warehouse = warehouse_from_params(&params)?;
-        let namespace = namespace_from_params(&params)?;
-        let table = table_name_from_params(&params)?;
-        let resource = TableCatalogResource::table(&warehouse, &namespace, &table);
-        authorize_table_catalog_resource_request(&req, &resource, AdminAction::SetTableLifecycleAction).await?;
-        ensure_table_bucket_enabled(&warehouse).await?;
-        let request = read_json_body::<crate::table_catalog::TableMaintenanceConfig>(req.input).await?;
-        let store = table_catalog_store()?;
-        let response = store
-            .put_table_maintenance_config(&warehouse, &namespace.public_name(), &table, request)
-            .await
-            .map_err(catalog_store_error)?;
-        build_json_response(StatusCode::OK, &response)
-    }
-}
-
-pub struct GetTableMaintenanceJobHandler {}
-
-#[async_trait::async_trait]
-impl Operation for GetTableMaintenanceJobHandler {
-    async fn call(&self, req: S3Request<Body>, params: Params<'_, '_>) -> S3Result<S3Response<(StatusCode, Body)>> {
-        let warehouse = warehouse_from_params(&params)?;
-        let namespace = namespace_from_params(&params)?;
-        let table = table_name_from_params(&params)?;
-        let job = job_id_from_params(&params)?;
-        let resource = TableCatalogResource::table(&warehouse, &namespace, &table);
-        authorize_table_catalog_resource_request(&req, &resource, AdminAction::GetTableLifecycleAction).await?;
-        ensure_table_bucket_enabled(&warehouse).await?;
-        let store = table_catalog_store()?;
-        let Some(response) = store
-            .get_table_metadata_maintenance_report(&warehouse, &namespace.public_name(), &table, &job)
-            .await
-            .map_err(catalog_store_error)?
-        else {
-            return Err(s3_error!(InvalidRequest, "maintenance job not found"));
-        };
-        build_json_response(StatusCode::OK, &response)
-    }
-}
-
-pub struct GetTableMaintenanceSchedulerHandler {}
-
-#[async_trait::async_trait]
-impl Operation for GetTableMaintenanceSchedulerHandler {
-    async fn call(&self, req: S3Request<Body>, params: Params<'_, '_>) -> S3Result<S3Response<(StatusCode, Body)>> {
-        let warehouse = warehouse_from_params(&params)?;
-        let namespace = namespace_from_params(&params)?;
-        let table = table_name_from_params(&params)?;
-        let resource = TableCatalogResource::table(&warehouse, &namespace, &table);
-        authorize_table_catalog_resource_request(&req, &resource, AdminAction::GetTableLifecycleAction).await?;
-        ensure_table_bucket_enabled(&warehouse).await?;
-        let store = table_catalog_store()?;
-        let response = store
-            .get_table_maintenance_scheduler_report(&warehouse, &namespace.public_name(), &table)
-            .await
-            .map_err(catalog_store_error)?;
-        build_json_response(StatusCode::OK, &response)
-    }
-}
-
-pub struct RunTableMaintenanceWorkerHandler {}
-
-#[async_trait::async_trait]
-impl Operation for RunTableMaintenanceSchedulerHandler {
-    async fn call(&self, req: S3Request<Body>, params: Params<'_, '_>) -> S3Result<S3Response<(StatusCode, Body)>> {
-        let warehouse = warehouse_from_params(&params)?;
-        let namespace = namespace_from_params(&params)?;
-        let table = table_name_from_params(&params)?;
-        let resource = TableCatalogResource::table(&warehouse, &namespace, &table);
-        authorize_table_catalog_resource_request(&req, &resource, AdminAction::RunTableMaintenanceAction).await?;
-        ensure_table_bucket_enabled(&warehouse).await?;
-        let request = read_json_body_or_default::<TableMaintenanceSchedulerRunRequest>(req.input).await?;
-        let store = table_catalog_store()?;
-        let response = store
-            .run_table_maintenance_scheduler_once(
-                &warehouse,
-                &namespace.public_name(),
-                &table,
-                request.scheduler_id().to_string(),
-            )
-            .await
-            .map_err(catalog_store_error)?;
-        build_json_response(StatusCode::OK, &response)
-    }
-}
-
-pub struct RunTableMaintenanceSchedulerHandler {}
-
-#[async_trait::async_trait]
-impl Operation for RunTableMaintenanceWorkerHandler {
-    async fn call(&self, req: S3Request<Body>, params: Params<'_, '_>) -> S3Result<S3Response<(StatusCode, Body)>> {
-        let warehouse = warehouse_from_params(&params)?;
-        let namespace = namespace_from_params(&params)?;
-        let table = table_name_from_params(&params)?;
-        let resource = TableCatalogResource::table(&warehouse, &namespace, &table);
-        authorize_table_catalog_resource_request(&req, &resource, AdminAction::RunTableMaintenanceAction).await?;
-        ensure_table_bucket_enabled(&warehouse).await?;
-        let request = read_json_body::<TableMaintenanceWorkerRunRequest>(req.input).await?;
-        let store = table_catalog_store()?;
-        let response = store
-            .run_table_metadata_maintenance_worker_once(
-                &warehouse,
-                &namespace.public_name(),
-                &table,
-                request.worker_id().to_string(),
-            )
-            .await
-            .map_err(catalog_store_error)?;
-        build_json_response(StatusCode::OK, &response)
-    }
-}
-
-pub struct HeartbeatTableMaintenanceJobHandler {}
-
-#[async_trait::async_trait]
-impl Operation for HeartbeatTableMaintenanceJobHandler {
-    async fn call(&self, req: S3Request<Body>, params: Params<'_, '_>) -> S3Result<S3Response<(StatusCode, Body)>> {
-        let warehouse = warehouse_from_params(&params)?;
-        let namespace = namespace_from_params(&params)?;
-        let table = table_name_from_params(&params)?;
-        let job = job_id_from_params(&params)?;
-        let resource = TableCatalogResource::table(&warehouse, &namespace, &table);
-        authorize_table_catalog_resource_request(&req, &resource, AdminAction::RunTableMaintenanceAction).await?;
-        ensure_table_bucket_enabled(&warehouse).await?;
-        let request = read_json_body::<TableMaintenanceHeartbeatRequest>(req.input).await?;
-        let store = table_catalog_store()?;
-        let response = store
-            .heartbeat_table_metadata_maintenance_job(
-                &warehouse,
-                &namespace.public_name(),
-                &table,
-                &job,
-                &request.lease_id,
-                &request.worker_id,
-            )
-            .await
-            .map_err(catalog_store_error)?;
-        build_json_response(StatusCode::OK, &response)
-    }
-}
-
-pub struct TableMaintenanceQuarantineHandler {}
-
-#[async_trait::async_trait]
-impl Operation for TableMaintenanceQuarantineHandler {
-    async fn call(&self, req: S3Request<Body>, params: Params<'_, '_>) -> S3Result<S3Response<(StatusCode, Body)>> {
-        let warehouse = warehouse_from_params(&params)?;
-        let namespace = namespace_from_params(&params)?;
-        let table = table_name_from_params(&params)?;
-        let job = job_id_from_params(&params)?;
-        let resource = TableCatalogResource::table(&warehouse, &namespace, &table);
-        authorize_table_catalog_resource_request(&req, &resource, AdminAction::RunTableMaintenanceAction).await?;
-        ensure_table_bucket_enabled(&warehouse).await?;
-        let request = read_json_body::<crate::table_catalog::TableMaintenanceQuarantineOperationRequest>(req.input).await?;
-        let store = table_catalog_store()?;
-        let response = store
-            .apply_table_maintenance_quarantine_operation(&warehouse, &namespace.public_name(), &table, &job, request)
-            .await
-            .map_err(catalog_store_error)?;
-        build_json_response(StatusCode::OK, &response)
-    }
-}
-
-pub struct ExportTableCatalogHandler {}
-
-#[async_trait::async_trait]
-impl Operation for ExportTableCatalogHandler {
-    async fn call(&self, req: S3Request<Body>, params: Params<'_, '_>) -> S3Result<S3Response<(StatusCode, Body)>> {
-        let warehouse = warehouse_from_params(&params)?;
-        let namespace = namespace_from_params(&params)?;
-        let table = table_name_from_params(&params)?;
-        let resource = TableCatalogResource::table(&warehouse, &namespace, &table);
-        authorize_table_catalog_resource_request(&req, &resource, AdminAction::GetTableMetadataAction).await?;
-        ensure_table_bucket_enabled(&warehouse).await?;
-        let store = table_catalog_store()?;
-        let started = Instant::now();
-        let result = store
-            .export_table_catalog_entry(&warehouse, &namespace.public_name(), &table)
-            .await
-            .map_err(catalog_store_error);
-        record_table_catalog_admin_operation_result("export", &warehouse, &namespace.public_name(), &table, started, &result);
-        let response = result?;
-        build_json_response(StatusCode::OK, &response)
-    }
-}
-
-pub struct ImportTableCatalogHandler {}
-
-#[async_trait::async_trait]
-impl Operation for ImportTableCatalogHandler {
-    async fn call(&self, req: S3Request<Body>, params: Params<'_, '_>) -> S3Result<S3Response<(StatusCode, Body)>> {
-        let warehouse = warehouse_from_params(&params)?;
-        let namespace = namespace_from_params(&params)?;
-        let table = table_name_from_params(&params)?;
-        let resource = TableCatalogResource::table(&warehouse, &namespace, &table);
-        authorize_table_catalog_resource_request(&req, &resource, AdminAction::RegisterTableAction).await?;
-        let request = read_json_body::<CatalogImportRequest>(req.input).await?;
-        let metadata_backend = table_catalog_backend()?;
-        let store = table_catalog_store_from_backend(metadata_backend.clone())?;
-        let table_bucket_enabled = table_bucket_enabled_from_metadata(&warehouse).await?;
-        let response =
-            catalog_import_response(&store, &metadata_backend, &warehouse, &namespace, &table, request, table_bucket_enabled)
-                .await?;
-        build_json_response(StatusCode::OK, &response)
-    }
-}
-
-pub struct ExternalCatalogBridgeHandler {}
-
-#[async_trait::async_trait]
-impl Operation for ExternalCatalogBridgeHandler {
-    async fn call(&self, req: S3Request<Body>, params: Params<'_, '_>) -> S3Result<S3Response<(StatusCode, Body)>> {
-        let warehouse = warehouse_from_params(&params)?;
-        let namespace = namespace_from_params(&params)?;
-        let table = table_name_from_params(&params)?;
-        let resource = TableCatalogResource::table(&warehouse, &namespace, &table);
-        authorize_table_catalog_resource_request(&req, &resource, AdminAction::GetTableMetadataAction).await?;
-        ensure_table_bucket_enabled(&warehouse).await?;
-        let store = table_catalog_object_store()?;
-        let response = external_catalog_bridge_response(&store, &warehouse, &namespace, &table).await?;
-        build_json_response(StatusCode::OK, &response)
-    }
-}
-
-pub struct PutExternalCatalogBridgeHandler {}
-
-#[async_trait::async_trait]
-impl Operation for PutExternalCatalogBridgeHandler {
-    async fn call(&self, req: S3Request<Body>, params: Params<'_, '_>) -> S3Result<S3Response<(StatusCode, Body)>> {
-        let warehouse = warehouse_from_params(&params)?;
-        let namespace = namespace_from_params(&params)?;
-        let table = table_name_from_params(&params)?;
-        let resource = TableCatalogResource::table(&warehouse, &namespace, &table);
-        authorize_table_catalog_resource_request(&req, &resource, AdminAction::RegisterTableAction).await?;
-        ensure_table_bucket_enabled(&warehouse).await?;
-        let request = read_json_body::<ExternalCatalogBridgeRequest>(req.input).await?;
-        let store = table_catalog_object_store()?;
-        let response = put_external_catalog_bridge_response(&store, &warehouse, &namespace, &table, request).await?;
-        build_json_response(StatusCode::OK, &response)
-    }
-}
-
-pub struct SyncExternalCatalogBridgeHandler {}
-
-#[async_trait::async_trait]
-impl Operation for SyncExternalCatalogBridgeHandler {
-    async fn call(&self, req: S3Request<Body>, params: Params<'_, '_>) -> S3Result<S3Response<(StatusCode, Body)>> {
-        let warehouse = warehouse_from_params(&params)?;
-        let namespace = namespace_from_params(&params)?;
-        let table = table_name_from_params(&params)?;
-        let resource = TableCatalogResource::table(&warehouse, &namespace, &table);
-        authorize_table_catalog_resource_request(&req, &resource, AdminAction::SetTableMetadataLocationAction).await?;
-        ensure_table_bucket_enabled(&warehouse).await?;
-        let metadata_backend = table_catalog_backend()?;
-        let store = table_catalog_object_store()?;
-        if store
-            .load_table(&warehouse, &namespace.public_name(), &table)
-            .await
-            .map_err(catalog_store_error)?
-            .is_none()
-        {
-            authorize_table_catalog_resource_request(&req, &resource, AdminAction::RegisterTableAction).await?;
-        }
-        let request = read_json_body::<ExternalCatalogBridgeSyncRequest>(req.input).await?;
-        let table_bucket_enabled = table_bucket_enabled_from_metadata(&warehouse).await?;
-        let response = sync_external_catalog_bridge_response(
-            &store,
-            &metadata_backend,
-            &warehouse,
-            &namespace,
-            &table,
-            request,
-            table_bucket_enabled,
-        )
-        .await?;
-        build_json_response(StatusCode::OK, &response)
-    }
-}
-
-pub struct GetTableCatalogDiagnosticsHandler {}
-
-#[async_trait::async_trait]
-impl Operation for GetTableCatalogDiagnosticsHandler {
-    async fn call(&self, req: S3Request<Body>, params: Params<'_, '_>) -> S3Result<S3Response<(StatusCode, Body)>> {
-        let warehouse = warehouse_from_params(&params)?;
-        let namespace = namespace_from_params(&params)?;
-        let table = table_name_from_params(&params)?;
-        let resource = TableCatalogResource::table(&warehouse, &namespace, &table);
-        authorize_table_catalog_resource_request(&req, &resource, AdminAction::GetTableMetadataAction).await?;
-        ensure_table_bucket_enabled(&warehouse).await?;
-        let store = table_catalog_store()?;
-        let config = store
-            .get_table_maintenance_config(&warehouse, &namespace.public_name(), &table)
-            .await
-            .map_err(catalog_store_error)?;
-        let started = Instant::now();
-        let result = store
-            .diagnose_table_catalog(&warehouse, &namespace.public_name(), &table, config.retain_recent_metadata_files)
-            .await
-            .map_err(catalog_store_error);
-        record_table_catalog_admin_operation_result(
-            "diagnostics",
-            &warehouse,
-            &namespace.public_name(),
-            &table,
-            started,
-            &result,
-        );
-        let response = result?;
-        build_json_response(StatusCode::OK, &response)
-    }
-}
-
-pub struct RecoverTableCatalogHandler {}
-
-#[async_trait::async_trait]
-impl Operation for RecoverTableCatalogHandler {
-    async fn call(&self, req: S3Request<Body>, params: Params<'_, '_>) -> S3Result<S3Response<(StatusCode, Body)>> {
-        let warehouse = warehouse_from_params(&params)?;
-        let namespace = namespace_from_params(&params)?;
-        let table = table_name_from_params(&params)?;
-        let resource = TableCatalogResource::table(&warehouse, &namespace, &table);
-        authorize_table_catalog_resource_request(&req, &resource, AdminAction::CommitTableAction).await?;
-        ensure_table_bucket_enabled(&warehouse).await?;
-        let store = table_catalog_store()?;
-        let started = Instant::now();
-        let result = store
-            .recover_table_commits(&warehouse, &namespace.public_name(), &table)
-            .await
-            .map_err(catalog_store_error);
-        record_table_catalog_admin_operation_result("recovery", &warehouse, &namespace.public_name(), &table, started, &result);
-        let response = result?;
-        build_json_response(StatusCode::OK, &response)
-    }
-}
-
-pub struct RollbackTableCatalogHandler {}
-
-#[async_trait::async_trait]
-impl Operation for RollbackTableCatalogHandler {
-    async fn call(&self, req: S3Request<Body>, params: Params<'_, '_>) -> S3Result<S3Response<(StatusCode, Body)>> {
-        let warehouse = warehouse_from_params(&params)?;
-        let namespace = namespace_from_params(&params)?;
-        let table = table_name_from_params(&params)?;
-        let resource = TableCatalogResource::table(&warehouse, &namespace, &table);
-        authorize_table_catalog_resource_request(&req, &resource, AdminAction::CommitTableAction).await?;
-        ensure_table_bucket_enabled(&warehouse).await?;
-        let request = read_json_body::<RollbackTableRequest>(req.input).await?;
-        let metadata_backend = table_catalog_backend()?;
-        let store = table_catalog_store_from_backend(metadata_backend.clone())?;
-        let response = rollback_table_response(&store, &metadata_backend, &warehouse, &namespace, &table, request).await?;
-        build_json_response(StatusCode::OK, &response)
-    }
 }
 
 #[cfg(test)]
